+++ visualization/app/codeCharta/state/effects/resetChosenMetrics/resetChosenMetrics.effect.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed } from "@angular/core/testing"
-import { metricDataSelector } from "../../selectors/accumulatedData/metricData/metricData.selector"
+import { nodeMetricDataSelector } from "../../../lenses/metrics/metricsLens.facade"
 import { areChosenMetricsAvailableSelector } from "../../selectors/allNecessaryRenderDataAvailable/areAllNecessaryRenderDataAvailable.selector"
 import { setAreaMetric } from "../../store/dynamicSettings/areaMetric/areaMetric.actions"
 import { setColorMetric } from "../../store/dynamicSettings/colorMetric/colorMetric.actions"
@@ -18,7 +18,7 @@ describe("resetChosenMetricsEffect", () => {
             providers: [
                 provideMockStore({
                     selectors: [
-                        { selector: metricDataSelector, value: { nodeMetricData: [] } },
+                        { selector: nodeMetricDataSelector, value: [] },
                         { selector: areChosenMetricsAvailableSelector, value: false }
                     ]
                 })
@@ -33,12 +33,10 @@ describe("resetChosenMetricsEffect", () => {
     })
 
     it("should apply matching metrics, when area, height and color metrics of matching category are available", () => {
-        store.overrideSelector(metricDataSelector, {
-            nodeMetricData: [
-                { name: "rloc", maxValue: 9001 },
-                { name: "mcc", maxValue: 9001 }
-            ]
-        } as any)
+        store.overrideSelector(nodeMetricDataSelector, [
+            { name: "rloc", maxValue: 9001 },
+            { name: "mcc", maxValue: 9001 }
+        ] as any)
         store.refreshState()
 
         expect(store.dispatch).toHaveBeenCalledTimes(4)
@@ -49,12 +47,10 @@ describe("resetChosenMetricsEffect", () => {
     })
 
     it("should apply available metrics when no matching scenario was found", () => {
-        store.overrideSelector(metricDataSelector, {
-            nodeMetricData: [
-                { name: "rloc", maxValue: 9001 },
-                { name: "loc", maxValue: 9001 }
-            ]
-        } as any)
+        store.overrideSelector(nodeMetricDataSelector, [
+            { name: "rloc", maxValue: 9001 },
+            { name: "loc", maxValue: 9001 }
+        ] as any)
         store.refreshState()
 
         expect(store.dispatch).toHaveBeenCalledTimes(4)
@@ -66,12 +62,10 @@ describe("resetChosenMetricsEffect", () => {
 
     it("should do nothing, when chosen metrics are still available", () => {
         store.overrideSelector(areChosenMetricsAvailableSelector, true)
-        store.overrideSelector(metricDataSelector, {
-            nodeMetricData: [
-                { name: "rloc", maxValue: 9001 },
-                { name: "loc", maxValue: 9001 }
-            ]
-        } as any)
+        store.overrideSelector(nodeMetricDataSelector, [
+            { name: "rloc", maxValue: 9001 },
+            { name: "loc", maxValue: 9001 }
+        ] as any)
         store.refreshState()
 
         expect(store.dispatch).not.toHaveBeenCalled()