+++ visualization/app/codeCharta/state/selectors/allNecessaryRenderDataAvailable/areAllNecessaryRenderDataAvailable.selector.ts
@@ -5,19 +5,18 @@ import { colorMetricSelector } from "../../store/dynamicSettings/colorMetric/col
 import { dynamicSettingsSelector } from "../../store/dynamicSettings/dynamicSettings.selector"
 import { heightMetricSelector } from "../../store/dynamicSettings/heightMetric/heightMetric.selector"
 import { filesSelector } from "../../../fileStore/store/files.selector"
-import { metricDataSelector } from "../accumulatedData/metricData/metricData.selector"
+import { nodeMetricDataSelector } from "../../../lenses/metrics/metricsLens.facade"
 import { areDynamicSettingsAvailable } from "./utils/areDynamicSettingsAvailable"
 import { areMetricsAvailable } from "./utils/areMetricsAvailable"
 
 const areFileStatesAvailableSelector = createSelector(filesSelector, files => fileStatesAvailable(files))
 
 export const areChosenMetricsAvailableSelector = createSelector(
-    metricDataSelector,
+    nodeMetricDataSelector,
     areaMetricSelector,
     colorMetricSelector,
     heightMetricSelector,
-    (metricData, areaMetric, colorMetric, heightMetric) =>
-        areMetricsAvailable(metricData.nodeMetricData, [areaMetric, colorMetric, heightMetric])
+    (nodeMetricData, areaMetric, colorMetric, heightMetric) => areMetricsAvailable(nodeMetricData, [areaMetric, colorMetric, heightMetric])
 )
 
 const areDynamicSettingsAvailableSelector = createSelector(dynamicSettingsSelector, dynamicSettings =>
@@ -25,12 +24,12 @@ const areDynamicSettingsAvailableSelector = createSelector(dynamicSettingsSelect
 )
 
 export const areAllNecessaryRenderDataAvailableSelector = createSelector(
-    metricDataSelector,
+    nodeMetricDataSelector,
     areFileStatesAvailableSelector,
     areChosenMetricsAvailableSelector,
     areDynamicSettingsAvailableSelector,
-    (metricData, areFileStatesAvailable, areChosenMetricsAvailable, areDynamicSettingsAvailable) => {
-        if (metricData.nodeMetricData === null || !areFileStatesAvailable || !areChosenMetricsAvailable || !areDynamicSettingsAvailable) {
+    (nodeMetricData, areFileStatesAvailable, areChosenMetricsAvailable, areDynamicSettingsAvailable) => {
+        if (nodeMetricData === null || !areFileStatesAvailable || !areChosenMetricsAvailable || !areDynamicSettingsAvailable) {
             return false
         }